Add better support for include/exclude filters
The include and exclude methods have been added to support an ordered
list of include and exclude patterns. The patterns are internally stored
and added to the rsync command in the order they were added.

The patterns method can be used to add multiple patterns through one
function call and is supported by the build method.

Fixes #2